com.eprovement.poptavka.shared.domain.type.OfferStateType removed. String changed to OfferStateType from enum package in OfferDetail class. FullOfferDetail used instead of ClientConversationDetail in UniversalTableWidget. Added creating offerStateImage column to UniversalAsynchTable.
- Nedalo sa to oddelit do viacerych komitov - urobil som jednu zmenu a uz sa to viezlo - musime vymazat stary kod, co sa nepouziva - vela referencii

git-svn-id: svn://46.137.95.166/poptavka@1128 61682b0c-635e-44ba-a85b-d27e7dc34134

+++ poptavka-web/src/main/java/com/eprovement/poptavka/server/service/admin/AdminRPCServiceImpl.java
@@ -12,6 +12,7 @@ import com.eprovement.poptavka.domain.user.Supplier;
 import com.eprovement.poptavka.domain.enums.Verification;
 import com.eprovement.poptavka.server.converter.Converter;
 import com.eprovement.poptavka.shared.domain.adminModule.ActivationEmailDetail;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -381,7 +382,7 @@ public class AdminRPCServiceImpl extends AutoinjectingRemoteService implements A
     public void updateOffer(OfferDetail offerDetail) {
         Offer offer = generalService.find(Offer.class, offerDetail.getId());
         if (!offer.getPrice().equals(offerDetail.getPrice())) {
-            offer.setPrice(offerDetail.getPrice());
+            offer.setPrice((BigDecimal) offerDetail.getPrice());
         }
         if (!offer.getCreated().equals(offerDetail.getCreatedDate())) {
             offer.setCreated(offerDetail.getCreatedDate());
@@ -395,7 +396,7 @@ public class AdminRPCServiceImpl extends AutoinjectingRemoteService implements A
         }
         //TODO Martin - how to update OfferState??
         if (!offer.getPrice().equals(offerDetail.getPrice())) {
-            offer.setPrice(offerDetail.getPrice());
+            offer.setPrice((BigDecimal) offerDetail.getPrice());
         }
         generalService.merge(offer);
     }
